PROJETO FINAL INFINITY/wayne-industries-requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
mysql-connector-python==8.0.26
python-multipart==0.0.5
